+++ app/src/main/java/com/team22/soundary/feature/search/presentation/friend/PendingFriendAdapter.kt
@@ -1,4 +1,4 @@
-package com.team22.soundary.feature.search
+package com.team22.soundary.feature.search.presentation.friend
 
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +8,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.team22.soundary.R
 import com.team22.soundary.core.domain.model.User
+import com.team22.soundary.feature.search.FriendDiffCallback
 
 class PendingFriendAdapter(
     private val onItemClick: (User) -> Unit
