+++ app/src/main/java/com/team22/soundary/feature/search/presentation/friend/SearchResultAdapter.kt
@@ -1,4 +1,4 @@
-package com.team22.soundary.feature.search
+package com.team22.soundary.feature.search.presentation.friend
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.team22.soundary.core.domain.model.User
 import com.team22.soundary.databinding.ItemSearchResultBinding
+import com.team22.soundary.feature.search.FriendDiffCallback
 
 class SearchResultAdapter(
     private val onRequestFriendClick: (User) -> Unit,
