+++ app/src/main/java/com/team22/soundary/feature/search/presentation/profile/FriendProfileFragment.kt
@@ -1,15 +1,13 @@
-package com.team22.soundary.feature.search
+package com.team22.soundary.feature.search.presentation.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import com.bumptech.glide.Glide
-import com.team22.soundary.core.domain.model.Category
 import com.team22.soundary.databinding.FragmentFriendProfileBinding
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
